Add one more haskell example and runnable links
I think the additional example is quite nice way to do the task in haskell, although the approach is a bit different compared to the others.

The website was the first I could find that would let you share the code with a link. It may nag about logging in when trying to run the snippet, but it is optional. If there is better alternative, the links could be changed.