trying to automate making changes and committing.
I have a clone on wsl. So I can use ninja to build and run tests. The Idea is if the standard tests pass or fail. the output should match known_good_test.txt. I was thinking of checking builds too but I think it wouldn't work. But the changes won't push automaticlly anymore. So I can push a batch of changes.