update python to check cmake and builds.
I was thinking that if build failed than tests would as well. Though it didn't seem to be the case so the script was commiting failed builds overnight. So i'll reset and try this today while I'm at work.

I made an array of commands and test.txt files and it runs the command checks the out put and returns true on failure.