Add numbool macro
This keeps the old behaviour of `num`, which means returning a boolean
for input [0, 1], and integers for other valid input values. We use this
to replace old calls to `num` that do (or may) depend on this behaviour.
This might be a temporary thing, at this moment we simply do not know
enough about the behaviour of the changed code.